MVC Menu updates
I refactored the CafeMenu controller so that it focuses primarily on handling the http request.

The filtering logic was moved to the BLL and a new method was built to handle this concern.

I added a new DTO to transfer the selected model properties, rather than sending the model into the logic layer.