Updates for order service, controller, and view
I added a new method to the order service that makes use of the recently implementation to the shopping bag repository. The controller was also updated to get the total for a customer's shopping bag.

The view now displays the customer's total before tip and features some sample tip amounts for guidance.